% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaglomr.R
\name{agglomerated}
\alias{agglomerated}
\title{Create an agglomerated dataframe.}
\usage{
agglomerated(features, samples, taxonomy, sample_col, otu_col = "otu_id",
  count_col = "count")
}
\arguments{
\item{features}{a matrix-like object with features as columns and samples as rows}

\item{samples}{an optional dataframe containing sample metadata, with each sample a separate row}

\item{taxonomy}{an optional dataframe containing taxonomy information, with each taxa a row and
rownames providing taxa id}

\item{sample_col}{a string giving the name of the column in `samples` corresponding to the rows
in the features table}

\item{otu_col}{a string giving the desired name of the otu column in the resulting dataframe}

\item{count_col}{a string giving the desired name of the counts column in the resulting dataframe}
}
\value{
A dataframe with each row representing a unique feature+sample+count observation, plus
  any metadata about the sample and any taxonomic annotation about the feature. This results in a
  lot of repetition, but is very easy to manipulate with dplyr and the tidyverse.
}
\description{
Creates an agglomerated dataframe from feature counts, sample metadata, and taxonomy annotations.
}
