% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaglomr.R
\name{agglomerated_to_phyloseq}
\alias{agglomerated_to_phyloseq}
\title{Convert an agglomerated dataframe to a phyloseq object.}
\usage{
agglomerated_to_phyloseq(agg, sample_col, otu_col, count_col, sample_metadata,
  tax_ranks)
}
\arguments{
\item{agg}{the agglomerated dataframe}

\item{sample_col}{the name of the sample id column in agg}

\item{otu_col}{the name of the otu id column in agg}

\item{count_col}{the unquoted name of the counts column in agg}

\item{sample_metadata}{unquoted names of sample metadata}

\item{tax_ranks}{unquoted names of taxonomic ranks (i.e. Kingdom:Species)}
}
\description{
Convert an agglomerated dataframe to a phyloseq object.
}
