% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaglomr.R
\docType{data}
\name{features}
\alias{features}
\title{Example feature counts.}
\format{An object of class \code{matrix} with 16 rows and 64 columns.}
\usage{
features
}
\description{
Example feature counts.
}
\keyword{datasets}
