% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaglomr.R
\name{get_features}
\alias{get_features}
\title{Extract feature count matrix from agglomerated dataframe}
\usage{
get_features(agg, sample_col, otu_col, count_col)
}
\arguments{
\item{agg}{agglomerated dataframe}

\item{sample_col}{unquoted name of the sample id column}

\item{otu_col}{unquoted name of the otu id column}

\item{count_col}{unquoted name of the count column}
}
\value{
the feature count matrix, invisibly
}
\description{
Extract feature count matrix from agglomerated dataframe
}
