% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaglomr.R
\name{get_taxa}
\alias{get_taxa}
\title{Extract taxonomic annotations from agglomerated dataframe}
\usage{
get_taxa(agg, otu_col, ...)
}
\arguments{
\item{agg}{agglomerated dataframe}

\item{otu_col}{unquoted name of the otu id column}

\item{...}{unquoted names of taxonomic ranks, i.e. Kingdom:Species}
}
\value{
a character matrix of taxonomic annotations
}
\description{
Extract taxonomic annotations from agglomerated dataframe
}
