% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaglomr.R
\name{phyloseq_to_agglomerated}
\alias{phyloseq_to_agglomerated}
\title{Convert a phyloseq object to an agglomerated (melted) dataframe.}
\usage{
phyloseq_to_agglomerated(ps, sample_col, otu_col, count_col)
}
\arguments{
\item{ps}{a phyloseq object}

\item{otu_col}{the desired name of the otu id column}

\item{count_col}{the desired name of the count column}

\item{sample.col.name}{the sample_data() column corresponding rownames(otu_table()), as a string}
}
\description{
Convert a phyloseq object to an agglomerated (melted) dataframe.
}
