% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaglomr.R
\docType{data}
\name{samples}
\alias{samples}
\title{Example sample metadata.}
\format{An object of class \code{data.frame} with 16 rows and 3 columns.}
\usage{
samples
}
\description{
Example sample metadata.
}
\keyword{datasets}
