% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaglomr.R
\docType{data}
\name{taxa}
\alias{taxa}
\title{Example taxonomy annotations}
\format{An object of class \code{data.frame} with 64 rows and 7 columns.}
\usage{
taxa
}
\description{
Example taxonomy annotations
}
\keyword{datasets}
