% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaglomr.R
\name{agglomerated_to_phyloseq}
\alias{agglomerated_to_phyloseq}
\title{Convert an agglomerated dataframe to a phyloseq object.}
\usage{
agglomerated_to_phyloseq(agg, otu_col, sample_col, count_col, ...)
}
\arguments{
\item{agg}{the agglomerated dataframe}

\item{otu_col}{the unquoted name of the otu id column in agg}

\item{sample_col}{the unquoted name of the sample id column in agg}

\item{count_col}{the unquoted name of the counts column in agg}

\item{...}{remaining unquoted names of taxonomic ranks (i.e. Kingdom:Species)}
}
\description{
Convert an agglomerated dataframe to a phyloseq object.
}
